requirements.txt
pycrypto==2.6.1
JPype1==0.7.5
JayDeBeApi==1.2.3
PyYAML==5.3.1
Sphinx==3.1.2
sphinxcontrib-websupport==1.2.3
openpyxl==3.0.4
psutil==5.7.2
regex==2020.7.14
requests==2.24.0
urllib3==1.25.9

pytest~=5.4.3
setuptools~=49.2.0